gift提交频率改为1，暂时解决超出整数范围的问题

+++ sql_const.py
@@ -94,7 +94,7 @@ privilege_type int,
 medal_level int,
 medal_name text,
 medal_room_id bigint,
-medal_room_uid int,
+medal_room_uid bigint,
 
 timestamp bigint,
 datatime timestamp
